requirements.txt
asgiref==3.4.1

certifi==2021.10.8
Django==4.0.1
djangorestframework==3.13.1
psycopg2==2.9.3
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
